Move the private read_signals helper out of sse.py
