+++ sdk/python/src/datastar_py/__init__.py
@@ -1,3 +1,22 @@
+from __future__ import annotations
+
+import json
+from typing import Mapping, Any
+
 from .sse import SSE_HEADERS, ServerSentEventGenerator
 
 __all__ = ["ServerSentEventGenerator", "SSE_HEADERS"]
+
+
+def _read_signals(
+    method: str, headers: Mapping, params: Mapping, body: str | bytes
+) -> dict[str, Any] | None:
+    if "Datastar-Request" not in headers:
+        return None
+    if method == "GET":
+        data = params.get("datastar")
+    elif headers.get("Content-Type") == "application/json":
+        data = body
+    else:
+        return None
+    return json.loads(data) if data else None
