+++ sdk/python/src/datastar_py/sse.py
@@ -1,9 +1,8 @@
 from __future__ import annotations
 
 import json
-from collections.abc import Mapping
 from itertools import chain
-from typing import Any, Optional, Protocol, Union, runtime_checkable
+from typing import Optional, Protocol, Union, runtime_checkable
 
 import datastar_py.consts as consts
 
@@ -178,17 +177,3 @@ class ServerSentEventGenerator:
     @classmethod
     def redirect(cls, location: str):
         return cls.execute_script(f"setTimeout(() => window.location = '{location}')")
-
-
-def _read_signals(
-    method: str, headers: Mapping, params: Mapping, body: str | bytes
-) -> dict[str, Any] | None:
-    if "Datastar-Request" not in headers:
-        return None
-    if method == "GET":
-        data = params.get("datastar")
-    elif headers.get("Content-Type") == "application/json":
-        data = body
-    else:
-        return None
-    return json.loads(data) if data else None
