+++ sdk/python/src/datastar_py/starlette.py
@@ -6,7 +6,8 @@ from typing import Any
 from starlette.requests import Request
 from starlette.responses import StreamingResponse as _StreamingResponse
 
-from .sse import SSE_HEADERS, ServerSentEventGenerator, _read_signals
+from .sse import SSE_HEADERS, ServerSentEventGenerator
+from . import _read_signals
 
 
 class DatastarStreamingResponse(_StreamingResponse):
